refactor: client코드와 server코드 분리
squash
