+++ src/constants.js
@@ -10,11 +10,3 @@ export const TMDB_MOVIE_LISTS = {
   upcoming: BASE_URL + "/upcoming?language=ko-KR&page=1",
 };
 export const TMDB_MOVIE_DETAIL_URL = "https://api.themoviedb.org/3/movie/";
-
-export const FETCH_OPTIONS = {
-  method: "GET",
-  headers: {
-    accept: "application/json",
-    Authorization: "Bearer " + process.env.TMDB_TOKEN,
-  },
-};
